# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aryventory Sales Management API",
        "description": "API for managing sales executives, leads, meetings, shifts, and business operations",
        "contact": {
            "email": "support@aryventory.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        },
        {
            "url": "https://api.aryventory.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/meetings/start": {
            "post": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Start a meeting",
                "description": "Start a meeting with selfie and location tracking",
                "operationId": "startMeeting",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "lead_id"
                                ],
                                "properties": {
                                    "lead_id": {
                                        "description": "Lead ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "selfie": {
                                        "description": "Selfie image",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "latitude": {
                                        "description": "Location latitude",
                                        "type": "number",
                                        "format": "float",
                                        "example": 28.6139,
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "description": "Location longitude",
                                        "type": "number",
                                        "format": "float",
                                        "example": 77.209,
                                        "nullable": true
                                    },
                                    "address": {
                                        "description": "Location address",
                                        "type": "string",
                                        "example": "123 Main Street, City",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Meeting started successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Meeting started successfully"
                                        },
                                        "meeting": {
                                            "description": "Meeting details",
                                            "type": "object"
                                        },
                                        "selfie_url": {
                                            "description": "Selfie image URL",
                                            "type": "string",
                                            "example": "https://s3.amazonaws.com/bucket/selfie.jpg",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "lead_id": [
                                                    "The lead id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": {
                "sanctum": []
            }
        }
    },
...
}